pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "comfy-gpu-offload"
version = "0.1.0"
description = "ComfyUI custom nodes for RunPod remote GPU execution."
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [{ name = "comfy-gpu-offload authors" }]
dependencies = [
    "requests>=2.32.0",
    "pillow>=12.0.0,<13.0.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.10.0",
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "types-requests>=2.32.0.20240907",
    "bandit>=1.7.9",
]

[dependency-groups]
dev = [
    "mypy>=1.10.0",
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "types-requests>=2.32.0.20240907",
    "bandit>=1.7.9",
]

[tool.ruff]
line-length = 100
target-version = "py313"
lint.select = [
    "E",
    "F",
    "B",  # bugbear
    "I",  # import order
    "UP", # pyupgrade
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.mypy]
python_version = "3.13"
check_untyped_defs = true
disallow_untyped_defs = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_optional = true
ignore_missing_imports = false
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]

